requirements.txt
psycopg2
kafka-python
requests
python-dotenv